DDP-7424: moved FieldSettings model into sub folder
