+++ ddp-workspace/projects/ddp-dsm-ui/src/app/tissue-list/tissue-list.component.ts
@@ -18,7 +18,7 @@ import { PatchUtil } from '../utils/patch.model';
 import { ModalComponent } from '../modal/modal.component';
 import { Utils } from '../utils/utils';
 import { TissueListWrapper } from './tissue-list-wrapper.model';
-import { FieldSettings } from '../field-settings/field-settings.model';
+import { FieldSettings } from '../field-settings/model/field-settings.model';
 import { Assignee } from '../assignee/assignee.model';
 import { AssigneeParticipant } from '../participant-list/models/assignee-participant.model';
